backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
python-multipart>=0.0.9
celery[redis]>=5.3.6
redis>=5.0.3
boto3>=1.34.0
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
